pyproject.toml
[tool.poetry]
name = "syntax-highligter-bot"
version = "0.1.0"
description = ""
authors = ["Muhammad Imangazaliev <imangazaliev.md@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.21.0"
aiogram = "^2.24"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
